Fixed issue with lapse rate, structure change
A raft of changes.

A lapse rate/ non-isothermal later issue was causing issues between ~32 km and ~47 km. The beta term (earth curvature) is removed completely.

Constants are now used, as well as the "Unitful". 

I have a fair degree of confidence in the output, but I'm sure the code will need to be tidied as I learn Julia.